Code cleanup, use the API to read and save cells, players
Rather than always just reading and saving cell and player data as raw
files, load cells with the API and only read player files when they
are offline.

I've also removed a bunch of unneeded code and comments.  The
`OnObjectSpawn` event just isn't needed, it's sufficient to simply zap
an already held unique if/when another player finds it.  It also
requires less code and hacky behavior; the previous method would
outright disable an entire container if it held a unique that was in
the DB.

And last but not least, the README has been updated to reflect new
options and to list default values where they had been missing.